Moved Event class from event.service.js to event.js
